+++ lib/fetures/dashboard/entities/screen_entities.dart
