+++ lib/fetures/property/property_listing_screen.dart
@@ -2,8 +2,7 @@ import 'dart:convert';
 
 import 'package:flutter/material.dart';
 import 'package:invoice_app/model/property.dart';
-import 'package:invoice_app/utils/extensions.dart';
-import 'package:invoice_app/widgets/properrty_card.dart';
+import 'package:invoice_app/widgets/property/properrty_card.dart';
 import 'package:loading_animation_widget/loading_animation_widget.dart';
 import 'package:supabase_flutter/supabase_flutter.dart';
 
@@ -16,19 +15,19 @@ class PropertyListingScreen extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return Scaffold(
-      appBar: AppBar(title: Text('Property')),
+      appBar: AppBar(title: const Text('Property')),
       body: FutureBuilder(
         future: fetch(),
         builder: (context, snapshot) {
           return switch (snapshot.connectionState) {
-            ConnectionState.none => Text('none'),
+            ConnectionState.none => const Text('none'),
             ConnectionState.waiting => Center(
               child: LoadingAnimationWidget.staggeredDotsWave(
                 color: Colors.white,
                 size: 80,
               ),
             ),
-            ConnectionState.active => Text('active'),
+            ConnectionState.active => const Text('active'),
             ConnectionState.done => RefreshIndicator(
               onRefresh: fetch,
               child: ListView.builder(
