+++ lib/fetures/settings/setting_screen.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:get_it/get_it.dart';
-import 'package:invoice_app/entities/auth_params/no_params.dart';
+import 'package:invoice_app/core/params/no_params.dart';
 import 'package:invoice_app/fetures/auth/login_screen.dart';
 import 'package:invoice_app/fetures/auth/widget/setting_tile.dart';
 import 'package:invoice_app/fetures/settings/riverpod/language_provider.dart';
