--- lib/utils/snackbar.dart
@@ -1,26 +1,0 @@
-import 'package:flutter/material.dart';
-
-class SnackbarHelper {
-  static void showSuccess(BuildContext context, String message) {
-    _showSnackbar(context, message, Colors.green);
-  }
-
-  static void showError(BuildContext context, String message) {
-    _showSnackbar(context, message, Colors.red);
-  }
-
-  static void _showSnackbar(
-    BuildContext context,
-    String message,
-    Color backgroundColor,
-  ) {
-    ScaffoldMessenger.of(context).showSnackBar(
-      SnackBar(
-        content: Text(message),
-        backgroundColor: backgroundColor,
-        behavior: SnackBarBehavior.floating,
-        duration: const Duration(seconds: 3),
-      ),
-    );
-  }
-}
