+++ lib/widgets/property/properrty_card.dart
